// types/RequestTypes.ts
export type RequestTypes = {
  id: string;
  title: string;
  startDay: string;
  startTime: string;
  endEnd: string;
  endTime: string;
  applicant: string;
  person: string;
  moreless: string;
  member: string[];
  level: string;
  content: string;
  display: boolean;
  deleteAt: boolean;
  editAt: boolean;
  sendAt: any;
  recruitment: boolean;
  author: string;
  endDay: string;
};